Improve publishers and their config (#70)
On macOS we got strange results, so try to improve LAN address detection. Also, add ability to configure. 

Defaults remain same: publisher opens a wildcard address (so on all interfaces) on a given port (or 0 for random ephemeral), and publisher uses (discovered or explicitly configured) LAN address and effective port to create URI.