Cache central only, but provide user config for list of cached reposes (#75)
So far Mimir handled Central only, on purpose. In version 0.6.0 there was a bold move, that is undone more properly: that version started handling all release-direct repositories.

Instead, provide config to allow users to tune which repo to cache. **By default Mimir is back to cache Central only**.

Also, contains a small bugfix: SNAPSHOT Mimir daemons should not join the cluster of RELEASE Mimir daemons, as if developer is using Mimir locally it usually leads to IT failures.