Global network hint (#83)
Mimir utilize multiple network accessing elements
(JGroups, Publishers), that on more complex setups with multiple interfaces and addresses may get
confused and bind to different addresses even.

Before it was possible to "hint" JGroups, but
Publisher may choose different address.

Now there is "global" hint, that is applied to
all networking elements, and ensures everything
is aligned -- bound to same interface and address.

Now it is enough to state in `~/.mimir/daemon.properties` this:
```
mimir.localHostHint=match-address\:192.168.1.*
```

And all of Mimir (JGroups, Publishers) will bind to this
address.