Cache Purge
Implement cache purge feature: ensures, that from file node
start (in daemon or embedded) to its end of life only those
cache entries remain in store that were used during lifespan.

Note: in IT we instruct daemon to stop on each invocation, hence
the one version we check for is present in cache only.