Cache Purge (#164)
Implement cache purge feature: to drop from cache artifacts never touched, to not "only grow" the caches, as on CI it is unwanted. Added two modes, "on begin" and "on end". The "on begin" is more aggressive, but whenever a build ends, it is guaranteed that cache contains only the touched artifacts. The "on end" is less eager, and also ensures, that from file node start (in daemon or embedded) to its end of life only those cache entries remain in store that were used during lifespan. This mode requires explicit daemon shut down.

## ON_BEGIN

In this mode whole store is moved to a "shadow" directory (so file node starts with _empty cache_), and cache entries are pulled back from "shadow" as needed. At file node close, the "shadow" directory is deleted and file node store directory will contain only those cached artifacts, that were used by any build or builds that were using this node.

## ON_END

In this mode the cache is kept "frozen" and a new, "shadow" directory is used by file node. On any key operation the frozen store is looked up and content, if any, is copied over to "shadow". At file node close, the "shadow" and file store are swapped out, hence at this point it is guaranteed that cache contains only those artifacts that were touched while file node existed.

Note: in IT we instruct daemon to stop on each invocation, hence the one version we check for is present in cache only.

Fixes #77