Bug: Daemon must reuse system node as local node
When Daemon pre-seeds, and creates MIMA/Resolver session,
it must reuse and not re-create local node, as if local
node is configured for exclusive access, it will fail.

So SessionConfig must have means to pass in live instance
of localNode. Also, when this happens, the session must
NOT attempt to close it, as the localNode is not managed
by it.

Added IT where Cache Purge happens combined with Pre-seed,
as this was the combination that revealed the issue
(as this is when Daemon creates MIMA/Resolver session).