Bug: Daemon must reuse system node as local node (#166)
When Daemon pre-seeds, and creates MIMA/Resolver session, it must reuse and not re-create local node, as if local node is configured for exclusive access, it will fail (by attempting to re-create new instance also having exclusive access). This makes combining cache-purge with pre-seed (where this happens) impossible.

Hence, SessionConfig must have means to pass in live instance of localNode. Also, when this happens, the session must NOT attempt to close it, as the localNode is not managed by it (as it is daemon systemNode).

Added IT where Cache Purge happens combined with Pre-seed, as this was the combination that revealed the issue (as this is when Daemon creates MIMA/Resolver session).