Full advisory locking (#167)
Implement full advisory locking, a side-struct which is used to coordinate, but do not lock the actual data we operate on, as some OS-es does not know to handle that.

This means that Mimir now does not lock data it operates on, to make possible combining of features like exclusive access and cache purge on certain operating systems.